app/backend/requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
requests==2.28.1
newsdataapi==0.1.21
Werkzeug == 2.3.6
openai==1.55.3 
datetime
python-dateutil == 2.8.2
newsapi-python